Use subdomain language for ingredients
