+++ src/index.js
@@ -7,7 +7,7 @@ import Questions from './Questions';
 import './app.css';
 
 axios.interceptors.response.use(res => {
-  console.log(res.data); // TODO: display ok
+  console.debug(res.data); // TODO: display ok
   return res;
 }, console.error); // TODO: display error
